+++ CHANGELOG.md
@@ -6,6 +6,9 @@
  AGENT_TRIGGER_METHOD_ON_AUTO_DEPLOY is set to incremental deployment
  - Fixed deadlock triggered by std::AgenConfigHandler (#1662)
 
+## Removed
+ - Removed the inmanta module validate command. Use pytest-inmanta fixtures to test your modules instead.
+
 # v 2019.5 (2019-12-05) Changes in this release:
 
 ## Fixed